server/requirements.txt
Flask==3.1.0
numpy==1.23.2
pandas==2.2.3
python-dotenv==1.0.1
scikit_learn==1.6.0
sentence_transformers==3.0.1
SQLAlchemy==2.0.36
